D-16: four unused components and the dead legacy configs move to obsolete/, nothing is deleted
The survey offered three options and the owner took none of them: everything goes to
`obsolete/`, and a class that happens to share a file with live code is moved *out* into its
own file rather than removed with it. Option (a) would have deleted the storage controller and
the `configuration.py` classes now and deferred the four components to the runtime-
SimRepository redesign, leaving them in limbo; option (c) would have removed the repository's
only MPC and PID prototypes for good. Moving all of it costs nothing and leaves the physics
and the parameter values findable.

Four components move with their four tests: `controller_mpc` (0 setup uses; 31 of its 49
config fields are runtime result buffers, which conversion would have forced into the file
format), `controller_pid` (0 setup uses), `generic_windturbine` (0 setup uses) and
`generic_price_signal` (0 setup uses, and superseded — `tariff_provider.py` publishes the same
two forecast keys from the same `TariffContract` the billing engine reads, so the parallel
phase they were both written for is over).

Two files gave up part of themselves. `SimpleHotWaterStorageController` and its config leave
`simple_water_storage.py` for `obsolete/components/simple_hot_water_storage_controller.py`;
the three live storage classes stay. That controller carried one of the two reads of
`WATERMASSFLOWRATEOFHEATGENERATOR`, a singleton key nothing writes any more, so one dead read
leaves `hisim/` here — the second, in `SimpleHotWaterStorage.__init__`, stays behind with the
live class and is not this commit's to remove. `WarmWaterStorageConfig`, `LoadConfig`,
`ElectricityDemandConfig` and `PVConfig` leave `configuration.py` for
`obsolete/components/configuration_legacy.py`, each verified by grep to have zero live
references first; `WarmWaterStorageConfig` is named only inside a commented-out block of
`generic_heat_pump.py`, which is left as the comment it is. Three neighbours deliberately
stay: `HouseholdWarmWaterDemandConfig` is **live** — the UTSP connector and
`simple_water_storage.py` read its class constants — and `HydrogenStorageConfig` and
`AdvElectrolyzerConfig` belong to D-25/D-29.

Two dead factories are deleted rather than moved, because a factory is code and not a class:
`RandomNumbersConfig.get_default_config` (the setups build the config directly, and both
scenario JSONs carry explicit configuration blocks, so `json_executor` never looks for it) and
`PVSystem.get_default_config` (a static method on the component returning a config, which
`json_executor` could not have used either, since it inspects the config class). Both were
confirmed to have zero callers across `hisim/`, `tests/`, `system_setups/`, `energy_systems/`
and `scripts/`.

`DeviceEnergySpecs.BY_CLASS_NAME` loses its `"MpcController"` and `"Windturbine"` rows and
`tests/test_economics_extraction.py` its two entries in the class-to-module table, because
that table's contract test requires every key to name a class that still lives in
`hisim.components`. There were no `PIDController` or `PriceSignal` rows to remove. The KPI
vocabulary is untouched and no recorded result moves: `Windturbine production` and its ratio
are keyed on `ComponentType.WINDTURBINE` in `loadtypes.py` and computed in
`kpi_preparation.py`, neither of which ever imported the retired module. `tariff_provider.py`
described the price signal and the MPC controller as live parallel-phase peers in four places;
those now read in the past tense and name `obsolete/` and this decision.

F-1 checked, and not triggered: none of the six retired classes has a
`get_default_connections_from_*` method on the energy management system. That component
declares exactly six, for the PV system, the UTSP occupancy, the more-advanced heat pump, the
electric heater, the battery and the solar thermal system, so no `add_component_output` call
disappears, no dynamic output is renumbered, and no scenario JSON goes stale — which is what
made D-1's blast radius twelve files.

Under D13 the zombie and defective thirds are now both closed. What is left is the data-
dependent remainder named in `plan.md`: `generic_smart_device.SmartDevice` (D-30), the H2/RSOC
classes whose only builders read two manufacturer JSONs absent from the repository (D-25,
D-26), and `GenericElectrolyzerConfig.get_default_config`, which takes a mandatory argument
(D-28, D-29).
