+++ hisim/components/configuration.py
@@ -2,11 +2,10 @@
 
 # clean
 
-from typing import Any, Optional
+from typing import Optional
 from dataclasses import dataclass, field
 from dataclasses_json import dataclass_json
 from hisim.loadtypes import LoadTypes, ComponentType
-from hisim.config import ConfigBase, ComponentID
 from hisim import log
 
 """
@@ -693,38 +692,6 @@ class EmissionFactorsAndCostsForDevicesConfig:
         return capex_techno_economic_values
 
 
-@dataclass_json
-@dataclass
-class WarmWaterStorageConfig(ConfigBase):
-    """Warm water storage config class."""
-
-    component_id: ComponentID
-    tank_diameter: float  # [m]
-    tank_height: float  # [m]
-    tank_start_temperature: float  # [°C]
-    temperature_difference: float  # [°C]
-    tank_u_value: float  # [W/m^2*K]
-    slice_height_minimum: float  # [m]
-
-    @classmethod
-    def get_default_config(
-        cls,
-        component_id: Optional[ComponentID] = None,
-    ) -> Any:
-        """Gets a default config."""
-        if component_id is None:
-            component_id = ComponentID(name="WarmWaterStorage")
-        return WarmWaterStorageConfig(
-            component_id=component_id,
-            tank_diameter=1,  # 0.9534        # [m]
-            tank_height=2,  # 3.15              # [m]
-            tank_start_temperature=65,  # [°C]
-            temperature_difference=0.3,  # [°C]
-            tank_u_value=0,  # 0.35                 # [W/m^2*K]
-            slice_height_minimum=0.05,  # [m]
-        )
-
-
 class GasHeaterConfig:
     """Gas heater config class."""
 
@@ -736,30 +703,6 @@ class GasHeaterConfig:
     delta_temperature = 25
     mass_flow_max = P_th_max / (4180 * delta_temperature)  # kg/s ## -> ~0.07
     temperature_max = 80  # [°C]
-
-
-class LoadConfig:
-    """Load config."""
-
-    # massflow_load_minute = 2.5          # [kg/min]
-    # massflow_load = massflow_load_minute / 60   # [kg/s]
-
-    possible_massflows_load = [0.1, 0.2, 0.3, 0.4]  # [kg/s]
-    delta_temperature = 20
-
-    # the returnflow shows if there was enough energy in the water
-    # -> use in load! Not in storage, there the water from WW is included
-    temperature_returnflow_minimum = 30  # [°C]
-
-    kwh_per_year = 20_201
-    demand_factor = kwh_per_year / 1000
-
-
-class ElectricityDemandConfig:
-    """Electricity demand config class."""
-
-    kwh_per_year = 6000
-    demand_factor = kwh_per_year / 1000
 
 
 class HouseholdWarmWaterDemandConfig:
@@ -824,12 +767,6 @@ class AdvElectrolyzerConfig:
 
     --> Solution: efficiency of the electrolyzer is calculated and is an Output
     """
-
-
-class PVConfig:
-    """PV config class."""
-
-    peak_power = 20_000  # [W]
 
 
 @dataclass_json
