+++ obsolete/components/generic_price_signal.py
