+++ obsolete/components/generic_windturbine.py
