+++ obsolete/tests/test_generic_price_signal.py
