+++ obsolete/tests/test_generic_windturbine.py
