+++ roadmap/declarative_energy_systems/p4_class_survey.md
@@ -997,6 +997,7 @@ The supplement proposes `standard` + `standard_5kwh`; the nominal factory pins *
 **D‑16 — The dead classes: delete or mint wire names for them?**
 `SimpleHotWaterStorageController(+Config)` (0 call sites at all), `MpcController` (0 setups; 31 of its 49 config fields are runtime buffers), `PIDController` (0 setups), `Windturbine` (0 setups), `PriceSignal` (0 setups), `RandomNumbersConfig.get_default_config` (0), `PVSystem.get_default_config` (0), and the five dead `configuration.py` classes. Options: (a) **Delete** the storage controller, the `configuration.py` five and the three zero‑call‑site factories now, and defer MPC/PID/wind/price to the runtime‑SimRepository redesign in the parking lot; (b) **convert everything**, minting ~9 preset names for code no setup runs; (c) delete all of them including the four components.
 *Consequence:* (a) removes the storage controller's copy of the dead singleton read for free and keeps the wire vocabulary honest, while leaving four components in limbo; (b) freezes nine names at P5 and forces `MpcControllerConfig`'s 31 result buffers into the file format; (c) removes the only MPC/PID prototypes in the repo.
+**Answered 2026-09-10, beyond the three options offered: everything to `obsolete/`, nothing deleted.** The four components move now with their tests rather than being deferred, and each class that shared a file (`SimpleHotWaterStorageController`, the four dead `configuration.py` classes) is moved out into its own file under `obsolete/components/`. `HouseholdWarmWaterDemandConfig` was found to be live and stays; `HydrogenStorageConfig` and `AdvElectrolyzerConfig` stay for D-25/D-29. Only the two dead factories are deleted — a factory is code, not a class.
 
 **D‑17 — Unify the two `PositionHotWaterStorageInSystemSetup` enums?**
 `simple_water_storage.py:58` (`PARALLEL_TO_HEAT_SOURCE`, `SERIES_TO_HEAT_SOURCE`) and `heat_distribution_system.py:61` (`PARALLEL`, `SERIES`, `NO_STORAGE_MASS_FLOW_FROM_HEAT_GENERATOR`, `NO_STORAGE_MASS_FLOW_FIX`) are different types with the same name, both are config fields, and both decide the same topology question — which is really a property of the *system*, not of either component (§6 lists it under topology queries). Options: (a) **Unify now**, before P5 freezes both spellings, into one enum in a shared module with the four‑member vocabulary; (b) leave both and let a file spell the same fact two ways; (c) derive both from a topology query in the executor and drop the fields (the §6 "presence/count" family, which the epic says stays out of the sizing engine).
