+++ tests/test_economics_extraction.py
@@ -1300,7 +1300,6 @@ _ENERGY_BALANCE_MODULES = {
     "FuelCellController": "controller_l1_fuel_cell",
     "L1Controller": "controller_l1_generic_ev_charge",
     "L1GenericElectrolyzerController": "controller_l1_electrolyzer",
-    "MpcController": "controller_mpc",
     "PTXController": "controller_l2_ptx_energy_management_system",
     "RsocBatteryController": "controller_l2_rsoc_battery_system",
     "XTPController": "controller_l2_xtp_fuel_cell_ems",
@@ -1308,7 +1307,6 @@ _ENERGY_BALANCE_MODULES = {
     "ComponentName": "example_template",
     "CarBattery": "advanced_ev_battery_bslib",
     "Car": "generic_car",
-    "Windturbine": "generic_windturbine",
     "ElectricHeating": "generic_electric_heating",
     "AirConditioner": "air_conditioner",
     "SimpleAirConditioner": "simple_air_conditioner",
